LayoutTests/fast/css/pseudo-active-on-labeled-element-not-canceled-by-focus-expected.txt

Verify that indirect :active on a labeled element is not affected when the element gets focus from the click.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



    Initial state
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
On Mouse Down
PASS elementsWithActiveStyle() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable", "sibling5"]
PASS elementsMatchingActiveSelector() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable"]
After Mouse Down
PASS elementsWithActiveStyle() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable", "sibling5"]
PASS elementsMatchingActiveSelector() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable"]
On Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
After Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
On Mouse Down
PASS elementsWithActiveStyle() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable", "sibling5"]
PASS elementsMatchingActiveSelector() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable"]
After Mouse Down
PASS elementsWithActiveStyle() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable", "sibling5"]
PASS elementsMatchingActiveSelector() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable"]
On Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
After Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
On Mouse Down
PASS elementsWithActiveStyle() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable", "sibling5"]
PASS elementsMatchingActiveSelector() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable"]
After Mouse Down
PASS elementsWithActiveStyle() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable", "sibling5"]
PASS elementsMatchingActiveSelector() is ["html", "body", "prime-ancestor", "labelable-ancestor", "labelable-parent", "target", "labelable"]
On Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
After Mouse Up
PASS elementsWithActiveStyle() is []
PASS elementsMatchingActiveSelector() is []
PASS successfullyParsed is true

TEST COMPLETE
